Update issue templates
Added/updated issue page on Curseforge & decided it would be a good policy to update the Github issue template as well. For now, keeping the issue pages separate, but may redirect Curseforge to Github in the future. For reference.. WowInterface does NOT appear to have an 'issue' page, but does have a comment page.